+++ src/frameworks/opengles/gles_guest.rs
@@ -74,6 +74,11 @@ fn glDisable(env: &mut Environment, cap: GLenum) {
         unsafe { gles.Disable(cap) };
     });
 }
+fn glClientActiveTexture(env: &mut Environment, texture: GLenum) {
+    with_ctx_and_mem(env, |gles, _mem| unsafe {
+        gles.ClientActiveTexture(texture)
+    })
+}
 fn glEnableClientState(env: &mut Environment, array: GLenum) {
     with_ctx_and_mem(env, |gles, _mem| {
         unsafe { gles.EnableClientState(array) };
@@ -116,6 +121,17 @@ fn glAlphaFuncx(env: &mut Environment, func: GLenum, ref_: GLclampx) {
 fn glBlendFunc(env: &mut Environment, sfactor: GLenum, dfactor: GLenum) {
     with_ctx_and_mem(env, |gles, _mem| unsafe {
         gles.BlendFunc(sfactor, dfactor)
+    })
+}
+fn glColorMask(
+    env: &mut Environment,
+    red: GLboolean,
+    green: GLboolean,
+    blue: GLboolean,
+    alpha: GLboolean,
+) {
+    with_ctx_and_mem(env, |gles, _mem| unsafe {
+        gles.ColorMask(red, green, blue, alpha)
     })
 }
 fn glCullFace(env: &mut Environment, mode: GLenum) {
@@ -161,6 +177,24 @@ fn glViewport(env: &mut Environment, x: GLint, y: GLint, width: GLsizei, height:
 }
 
 // Lighting and materials
+fn glFogf(env: &mut Environment, pname: GLenum, param: GLfloat) {
+    with_ctx_and_mem(env, |gles, _mem| unsafe { gles.Fogf(pname, param) })
+}
+fn glFogx(env: &mut Environment, pname: GLenum, param: GLfixed) {
+    with_ctx_and_mem(env, |gles, _mem| unsafe { gles.Fogx(pname, param) })
+}
+fn glFogfv(env: &mut Environment, pname: GLenum, params: ConstPtr<GLfloat>) {
+    with_ctx_and_mem(env, |gles, mem| {
+        let params = mem.ptr_at(params, 4 /* upper bound */);
+        unsafe { gles.Fogfv(pname, params) }
+    })
+}
+fn glFogxv(env: &mut Environment, pname: GLenum, params: ConstPtr<GLfixed>) {
+    with_ctx_and_mem(env, |gles, mem| {
+        let params = mem.ptr_at(params, 4 /* upper bound */);
+        unsafe { gles.Fogxv(pname, params) }
+    })
+}
 fn glLightf(env: &mut Environment, light: GLenum, pname: GLenum, param: GLfloat) {
     with_ctx_and_mem(env, |gles, _mem| unsafe {
         gles.Lightf(light, pname, param)
@@ -616,7 +650,34 @@ fn glTexSubImage2D(
         };
 
         gles.TexSubImage2D(
-            target, level, xoffset, yoffset, width, height, format, type_, pixels,
+            target, level, xoffset, yoffset, width, height, format, type_, pixels
+        )
+    })
+}
+fn glCompressedTexImage2D(
+    env: &mut Environment,
+    target: GLenum,
+    level: GLint,
+    internalformat: GLenum,
+    width: GLsizei,
+    height: GLsizei,
+    border: GLint,
+    image_size: GLsizei,
+    data: ConstVoidPtr,
+) {
+    with_ctx_and_mem(env, |gles, mem| unsafe {
+        let data = mem
+            .ptr_at(data.cast::<u8>(), image_size.try_into().unwrap())
+            .cast();
+        gles.CompressedTexImage2D(
+            target,
+            level,
+            internalformat,
+            width,
+            height,
+            border,
+            image_size,
+            data
         )
     })
 }
@@ -812,6 +873,7 @@ pub const FUNCTIONS: FunctionExports = &[
     export_c_func!(glGetError()),
     export_c_func!(glEnable(_)),
     export_c_func!(glDisable(_)),
+    export_c_func!(glClientActiveTexture(_)),
     export_c_func!(glEnableClientState(_)),
     export_c_func!(glDisableClientState(_)),
     export_c_func!(glGetBooleanv(_, _)),
@@ -822,6 +884,7 @@ pub const FUNCTIONS: FunctionExports = &[
     export_c_func!(glAlphaFunc(_, _)),
     export_c_func!(glAlphaFuncx(_, _)),
     export_c_func!(glBlendFunc(_, _)),
+    export_c_func!(glColorMask(_, _, _, _)),
     export_c_func!(glCullFace(_)),
     export_c_func!(glDepthFunc(_)),
     export_c_func!(glDepthMask(_)),
@@ -832,6 +895,10 @@ pub const FUNCTIONS: FunctionExports = &[
     export_c_func!(glScissor(_, _, _, _)),
     export_c_func!(glViewport(_, _, _, _)),
     // Lighting and materials
+    export_c_func!(glFogf(_, _)),
+    export_c_func!(glFogx(_, _)),
+    export_c_func!(glFogfv(_, _)),
+    export_c_func!(glFogxv(_, _)),
     export_c_func!(glLightf(_, _, _)),
     export_c_func!(glLightx(_, _, _)),
     export_c_func!(glLightfv(_, _, _)),
@@ -895,6 +962,7 @@ pub const FUNCTIONS: FunctionExports = &[
     export_c_func!(glTexParameteriv(_, _, _)),
     export_c_func!(glTexImage2D(_, _, _, _, _, _, _, _, _)),
     export_c_func!(glTexSubImage2D(_, _, _, _, _, _, _, _, _)),
+    export_c_func!(glCompressedTexImage2D(_, _, _, _, _, _, _, _)),
     export_c_func!(glCopyTexImage2D(_, _, _, _, _, _, _, _)),
     export_c_func!(glCopyTexSubImage2D(_, _, _, _, _, _, _, _)),
     export_c_func!(glTexEnvf(_, _, _)),