+++ src/window/gl_bindings/build.rs
@@ -24,6 +24,7 @@ fn main() {
         [
             "GL_EXT_framebuffer_object",
             "GL_EXT_texture_filter_anisotropic",
+            "GL_EXT_texture_lod_bias",
         ],
     )
     .write_bindings(GlobalGenerator, &mut file)
@@ -39,6 +40,10 @@ fn main() {
             "GL_OES_framebuffer_object",
             "GL_OES_rgb8_rgba8",
             "GL_EXT_texture_filter_anisotropic",
+            "GL_IMG_texture_compression_pvrtc",
+            "GL_EXT_texture_lod_bias",
+            // Part of the OpenGL ES 1.1 common profile.
+            "GL_OES_compressed_paletted_texture",
         ],
     )
     .write_bindings(GlobalGenerator, &mut file)